serverless/fauna/requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
faunadb==4.3.0
future==0.18.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
iso8601==1.1.0
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
